Pipeline/requirements.txt
astropy==3.1.1
attrs==17.3.0
backcall==0.1.0
bleach==2.1.4
cycler==0.10.0
Cython==0.25.2
decorator==4.3.0
defusedxml==0.5.0
Django==2.1.5
easygui==0.98.1
entrypoints==0.2.3
html5lib==1.0.1
image==1.5.27
ipykernel==4.9.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.1
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==2.2.3
mistune==0.8.3
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.6.0
numpy==1.16.0
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.4.1
pluggy==0.6.0
prometheus-client==0.3.1
prompt-toolkit==1.0.15
ptyprocess==0.6.0
py==1.6.0
Pygments==2.2.0
pyparsing==2.2.0
pytest==3.3.1
python-dateutil==2.7.3
pytz==2018.5
pyzmq==17.1.2
qtconsole==4.4.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.10.0
terminado==0.8.1
testpath==0.3.1
tornado==5.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.1
